+++ src/ll_movement/util/BoundedDisplacementCalculator.java
@@ -1,27 +1,29 @@
-package movement.util;
+package ll_movement.util;
 
 /**
- * Object that calculates position based on elapsed time from a velocity curve defined by displacement, max velocity, and max acceleration.
+ * Object that calculates position based on elapsed time from a velocity curve defined by displacement, time, max velocity, and max acceleration.
  */
-public class DisplacementCalculator {
+public class BoundedDisplacementCalculator {
 
 	private double distance, time, sign;
-	private final double MV, MA;
-	
+	private double MV, MA;
+
 	/**
-	 * Creates a new DisplacementCalculator with a given target and kinematic constraints.
+	 * Creates a new BoundedDisplacementCalculator with a given target, time, and kinematic constraints.
 	 * @param targetDisplacement
+	 * @param targetTime
 	 * @param MV
 	 * @param MA
 	 */
-	public DisplacementCalculator(double targetDisplacement, double MV, double MA) {
+	public BoundedDisplacementCalculator(double targetDisplacement, double targetTime, double MV, double MA) {
 		this.sign = Math.signum(targetDisplacement);
 		this.distance = Math.abs(targetDisplacement);
+		this.time = targetTime;
 		this.MV = MV;
 		this.MA = MA;
 		init();
 	}
-	
+
 	/**
 	 * @return the absolute value of the target displacement.
 	 */
@@ -35,49 +37,11 @@ public class DisplacementCalculator {
 	public double getTotalDisplacement() {
 		return distance * sign;
 	}
-	
+
 	/**
 	 * @return the minimum time needed to reach the target displacement value.
 	 */
 	public double getTime() {
-		return time;
-	}
-	
-	/**
-	 * Calculates the elapsed time at a certain displacement value.
-	 * @param displacement signed displacement
-	 * @return the elapsed time at which the calculator has reached the given displacement value.
-	 */
-	public double getElapsedTime(double displacement) {
-		displacement = bound(displacement * sign, 0, distance); // positive in [0, distance]
-		
-		double time;
-		double d_n = distance - displacement, d_a = 0.5 * MV*MV/MA, t_a = MV/MA;
-		if (this.time <= 2*t_a) {
-			// triangle graph
-			if (displacement <= distance/2)
-				time = Math.sqrt(2*displacement/MA);
-			else
-				time = this.time - Math.sqrt(2*d_n/MA);
-		} 
-		else {
-			// trapezoid graph
-			if (displacement <= distance/2)
-				if (displacement <= d_a)
-					// in first slope
-					time = Math.sqrt(2*displacement/MA);
-				else
-					// in first plateau
-					time = displacement/MV + t_a/2d;
-			else
-				if (d_n <= d_a)
-					// in last slope
-					time = this.time - Math.sqrt(2*d_n/MA);
-				else
-					// in second plateau
-					time = this.time - (d_n/MV + t_a/2d);
-		}
-		
 		return time;
 	}
 
@@ -89,20 +53,23 @@ public class DisplacementCalculator {
 	public double getDisplacement(double elapsedTime) {
 		elapsedTime = bound(elapsedTime, 0, time);
 		
+		double D = Math.abs(this.distance);
 		double displacement;
+		
 		double t_n = time - elapsedTime, t_a = MV/MA;
 		if (time <= 2*t_a) {
 			// triangle graph
 			if (elapsedTime <= time/2)
 				displacement = 0.5*MA*elapsedTime*elapsedTime;
 			else
-				displacement = distance - 0.5*MA*t_n*t_n;
-		} else {
+				displacement = D - 0.5*MA*t_n*t_n;
+		} 
+		else {
 			// trapezoid graph
 			if (elapsedTime <= time/2)
 				displacement = 0.5*(elapsedTime + Math.max(0, elapsedTime - t_a))* Math.min(MV, MA*elapsedTime);
 			else
-				displacement = distance - 0.5*(t_n + Math.max(0, t_n - t_a))* Math.min(MV, MA*t_n);
+				displacement = D - 0.5*(t_n + Math.max(0, t_n - t_a))* Math.min(MV, MA*t_n);
 		}
 		
 		displacement *= sign;
@@ -116,9 +83,11 @@ public class DisplacementCalculator {
 	 * @return the velocity value the given elapsed time.
 	 */
 	public double getVelocity(double elapsedTime) {
+		if (distance == 0) return 0;
 		elapsedTime = bound(elapsedTime, 0, time);
 		
 		double velocity;
+		
 		double t_n = time - elapsedTime, t_a = MV/MA;
 		if (time <= 2*t_a) {
 			// triangle graph
@@ -126,7 +95,8 @@ public class DisplacementCalculator {
 				velocity = MA*elapsedTime;
 			else
 				velocity = MA*t_n;
-		} else {
+		} 
+		else {
 			// trapezoid graph
 			if (elapsedTime <= time/2)
 				velocity = Math.min(MV, MA*elapsedTime);
@@ -138,23 +108,41 @@ public class DisplacementCalculator {
 		
 		return velocity;
 	}
-	
+
 	/**
-	 * Calculates total time.
+	 * Calculates max distance, min time, and max velocity.
 	 */
-	private void init() {
+	public void init() {
 		
-		double d_a = 0.5 * MV*MV / MA;
+		double T = time;
+		double t_a = MV/MA;
+		double D_max;
 		
-		if (distance / 2 <= d_a) {
+		if (T < 2*t_a) {
 			// triangle graph
-			time = 2*Math.sqrt(distance/MA);
-		} else {
+			D_max = MA * T * T / 4.0;
+		} 
+		else {
 			// trapezoid graph
-			time = distance/MV + MV/MA;
+			D_max = (T - MV/MA) * MV;
 		}
+		
+		if (distance >= D_max) {
+			// no need to compensate MV
+			distance = D_max;
+		}
+
+		// trim time
+		double previousMV = MV;
+		MV = Math.min(MV, Math.sqrt(distance*MA));
+		
+		if (Math.abs(MV-previousMV)<1e-3)
+			time = distance/MV + t_a;
+		else
+			time = Math.sqrt(4*distance/MA);
+		
 	}
-	
+
 	/**
 	 * Clips the input x between a given lower and upper bound.
 	 * @param x
