+++ src/ll_movement/util/DisplacementCalculator.java
@@ -1,29 +1,27 @@
-package movement.util;
+package ll_movement.util;
 
 /**
- * Object that calculates position based on elapsed time from a velocity curve defined by displacement, time, max velocity, and max acceleration.
+ * Object that calculates position based on elapsed time from a velocity curve defined by displacement, max velocity, and max acceleration.
  */
-public class BoundedDisplacementCalculator {
+public class DisplacementCalculator {
 
 	private double distance, time, sign;
-	private double MV, MA;
-
+	private final double MV, MA;
+	
 	/**
-	 * Creates a new BoundedDisplacementCalculator with a given target, time, and kinematic constraints.
+	 * Creates a new DisplacementCalculator with a given target and kinematic constraints.
 	 * @param targetDisplacement
-	 * @param targetTime
 	 * @param MV
 	 * @param MA
 	 */
-	public BoundedDisplacementCalculator(double targetDisplacement, double targetTime, double MV, double MA) {
+	public DisplacementCalculator(double targetDisplacement, double MV, double MA) {
 		this.sign = Math.signum(targetDisplacement);
 		this.distance = Math.abs(targetDisplacement);
-		this.time = targetTime;
 		this.MV = MV;
 		this.MA = MA;
 		init();
 	}
-
+	
 	/**
 	 * @return the absolute value of the target displacement.
 	 */
@@ -37,11 +35,49 @@ public class BoundedDisplacementCalculator {
 	public double getTotalDisplacement() {
 		return distance * sign;
 	}
-
+	
 	/**
 	 * @return the minimum time needed to reach the target displacement value.
 	 */
 	public double getTime() {
+		return time;
+	}
+	
+	/**
+	 * Calculates the elapsed time at a certain displacement value.
+	 * @param displacement signed displacement
+	 * @return the elapsed time at which the calculator has reached the given displacement value.
+	 */
+	public double getElapsedTime(double displacement) {
+		displacement = bound(displacement * sign, 0, distance); // positive in [0, distance]
+		
+		double time;
+		double d_n = distance - displacement, d_a = 0.5 * MV*MV/MA, t_a = MV/MA;
+		if (this.time <= 2*t_a) {
+			// triangle graph
+			if (displacement <= distance/2)
+				time = Math.sqrt(2*displacement/MA);
+			else
+				time = this.time - Math.sqrt(2*d_n/MA);
+		} 
+		else {
+			// trapezoid graph
+			if (displacement <= distance/2)
+				if (displacement <= d_a)
+					// in first slope
+					time = Math.sqrt(2*displacement/MA);
+				else
+					// in first plateau
+					time = displacement/MV + t_a/2d;
+			else
+				if (d_n <= d_a)
+					// in last slope
+					time = this.time - Math.sqrt(2*d_n/MA);
+				else
+					// in second plateau
+					time = this.time - (d_n/MV + t_a/2d);
+		}
+		
 		return time;
 	}
 
@@ -53,23 +89,20 @@ public class BoundedDisplacementCalculator {
 	public double getDisplacement(double elapsedTime) {
 		elapsedTime = bound(elapsedTime, 0, time);
 		
-		double D = Math.abs(this.distance);
 		double displacement;
-		
 		double t_n = time - elapsedTime, t_a = MV/MA;
 		if (time <= 2*t_a) {
 			// triangle graph
 			if (elapsedTime <= time/2)
 				displacement = 0.5*MA*elapsedTime*elapsedTime;
 			else
-				displacement = D - 0.5*MA*t_n*t_n;
-		} 
-		else {
+				displacement = distance - 0.5*MA*t_n*t_n;
+		} else {
 			// trapezoid graph
 			if (elapsedTime <= time/2)
 				displacement = 0.5*(elapsedTime + Math.max(0, elapsedTime - t_a))* Math.min(MV, MA*elapsedTime);
 			else
-				displacement = D - 0.5*(t_n + Math.max(0, t_n - t_a))* Math.min(MV, MA*t_n);
+				displacement = distance - 0.5*(t_n + Math.max(0, t_n - t_a))* Math.min(MV, MA*t_n);
 		}
 		
 		displacement *= sign;
@@ -83,11 +116,9 @@ public class BoundedDisplacementCalculator {
 	 * @return the velocity value the given elapsed time.
 	 */
 	public double getVelocity(double elapsedTime) {
-		if (distance == 0) return 0;
 		elapsedTime = bound(elapsedTime, 0, time);
 		
 		double velocity;
-		
 		double t_n = time - elapsedTime, t_a = MV/MA;
 		if (time <= 2*t_a) {
 			// triangle graph
@@ -95,8 +126,7 @@ public class BoundedDisplacementCalculator {
 				velocity = MA*elapsedTime;
 			else
 				velocity = MA*t_n;
-		} 
-		else {
+		} else {
 			// trapezoid graph
 			if (elapsedTime <= time/2)
 				velocity = Math.min(MV, MA*elapsedTime);
@@ -108,41 +138,23 @@ public class BoundedDisplacementCalculator {
 		
 		return velocity;
 	}
-
+	
 	/**
-	 * Calculates max distance, min time, and max velocity.
+	 * Calculates total time.
 	 */
-	public void init() {
+	private void init() {
 		
-		double T = time;
-		double t_a = MV/MA;
-		double D_max;
+		double d_a = 0.5 * MV*MV / MA;
 		
-		if (T < 2*t_a) {
+		if (distance / 2 <= d_a) {
 			// triangle graph
-			D_max = MA * T * T / 4.0;
-		} 
-		else {
+			time = 2*Math.sqrt(distance/MA);
+		} else {
 			// trapezoid graph
-			D_max = (T - MV/MA) * MV;
+			time = distance/MV + MV/MA;
 		}
-		
-		if (distance >= D_max) {
-			// no need to compensate MV
-			distance = D_max;
-		}
-
-		// trim time
-		double previousMV = MV;
-		MV = Math.min(MV, Math.sqrt(distance*MA));
-		
-		if (Math.abs(MV-previousMV)<1e-3)
-			time = distance/MV + t_a;
-		else
-			time = Math.sqrt(4*distance/MA);
-		
 	}
-
+	
 	/**
 	 * Clips the input x between a given lower and upper bound.
 	 * @param x
