+++ src/ll_movement/util/MovementType.java
@@ -1,4 +1,4 @@
-package movement.util;
+package ll_movement.util;
 
 /**
  * The type of Movement.
